feat: refactor post creation and update logic, improve error handling, and reorganize cloudinary utility
