+++ src/middlewares/upload.js
@@ -1,6 +1,6 @@
 import multer from 'multer'
 import { CloudinaryStorage } from 'multer-storage-cloudinary'
-import cloudinary from '../../cloudinary.js'
+import cloudinary from '../utils/cloudinary.js'
 
 const storage = new CloudinaryStorage({
   cloudinary,
