+++ src/schemas/posts.schema.js
@@ -4,8 +4,6 @@ export const createPostSchema = z.object({
   title: z.string().min(1, { message: 'El título es requerido' }),
   content: z.string().min(1, { message: 'El contenido es requerido' }),
   category_id: z.coerce.number({ message: 'La categoría es requerida' }),
-  community_id: z.coerce.number().optional(),
-  status: z.enum(['draft', 'published', 'pending_approval']).optional(), // Hacer que sea opcional
 })
 
 export const updatePostSchema = createPostSchema.partial()
