--- src/schemas/postsCategories.schema.js
@@ -1,7 +1,0 @@
-import { z } from 'zod'
-
-export const createPostCategorySchema = z.object({
-  name: z.enum(['Project', 'Event', 'News', 'Announcement'], {
-    required_error: 'El nombre de la categoría es requerido',
-  }),
-})
