+++ src/utils/cloudinary.js
@@ -3,6 +3,9 @@ import dotenv from 'dotenv'
 dotenv.config()
 
 cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
   secure: true,
 })
